flake: stop following zynq-rs, update rust
Cargo-xbuild (xargo) is essentially deprecated now that build-std is supported in rust-proper. Unfortunately, cargo doesn't properly find the rust source unless it's installed using rustup, which manages the toolchain as well, effectively rendering the flake-defined rust platform moot. I'm going to leave the builds there in the hope that will be fixed by the time I'm ready to deploy.